requirements.txt
flask==3.1.0
flask-sqlalchemy==3.1.1
flask-jwt-extended==4.7.1
mysql==0.0.3
mysql-connector-python==9.2.0
pymysql==1.1.1
pymongo==4.12.0
python-dotenv==1.1.0
bcrypt==4.3.0
sendgrid==6.11.0
google-auth==2.39.0